requirements.txt
-i https://pypi.org/simple
aiofiles==24.1.0; python_version >= '3.8'
aiohappyeyeballs==2.6.1; python_version >= '3.9'
aiohttp==3.12.6; python_version >= '3.9'
aiosignal==1.3.2; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
asgiref==3.8.1; python_version >= '3.8'
attrs==25.3.0; python_version >= '3.8'
backoff==2.2.1; python_version >= '3.7' and python_version < '4.0'
bcrypt==4.3.0; python_version >= '3.8'
build==1.2.2.post1; python_version >= '3.8'
cachetools==5.5.2; python_version >= '3.7'
certifi==2025.4.26; python_version >= '3.6'
cffi==1.17.1; platform_python_implementation != 'PyPy'
charset-normalizer==3.4.2; python_version >= '3.7'
chroma-hnswlib==0.7.3
chromadb==0.4.22; python_version >= '3.8'
click==8.2.1; python_version >= '3.10'
coloredlogs==15.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
couchdb==1.2
cryptography==45.0.3; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
dataclasses-json==0.6.7; python_version >= '3.7' and python_version < '4.0'
deprecated==1.2.18; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
distro==1.9.0; python_version >= '3.6'
dnspython==2.7.0; python_version >= '3.9'
durationpy==0.10
email-validator==2.2.0; python_version >= '3.8'
fastapi==0.115.12; python_version >= '3.8'
ffmpy==0.5.0; python_version >= '3.8' and python_version < '4.0'
filelock==3.18.0; python_version >= '3.9'
flatbuffers==25.2.10
frozenlist==1.6.0; python_version >= '3.9'
fsspec==2025.5.1; python_version >= '3.9'
google-auth==2.40.2; python_version >= '3.7'
googleapis-common-protos==1.70.0; python_version >= '3.7'
gradio==5.32.0; python_version >= '3.10'
gradio-client==1.10.2; python_version >= '3.10'
groovy==0.1.2; python_version >= '3.10'
grpcio==1.71.0; python_version >= '3.9'
h11==0.16.0; python_version >= '3.8'
hf-xet==1.1.2; platform_machine == 'x86_64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'aarch64'
httpcore==1.0.9; python_version >= '3.8'
httptools==0.6.4
httpx==0.28.1; python_version >= '3.8'
huggingface-hub==0.32.3; python_full_version >= '3.8.0'
humanfriendly==10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.10; python_version >= '3.6'
importlib-metadata==8.6.1; python_version >= '3.9'
importlib-resources==6.5.2; python_version >= '3.9'
jinja2==3.1.6; python_version >= '3.7'
jiter==0.10.0; python_version >= '3.9'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
kubernetes==32.0.1; python_version >= '3.6'
langchain==0.1.0; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-community==0.0.20; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-core==0.1.23; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-openai==0.0.6; python_version < '4.0' and python_full_version >= '3.8.1'
langsmith==0.0.87; python_version < '4.0' and python_full_version >= '3.8.1'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
marshmallow==3.26.1; python_version >= '3.9'
mdurl==0.1.2; python_version >= '3.7'
mmh3==5.1.0; python_version >= '3.9'
mpmath==1.3.0
multidict==6.4.4; python_version >= '3.9'
mypy-extensions==1.1.0; python_version >= '3.8'
numpy==1.26.4; python_version >= '3.9'
oauthlib==3.2.2; python_version >= '3.6'
onnxruntime==1.22.0; python_version >= '3.10'
openai==1.82.1; python_version >= '3.8'
opentelemetry-api==1.33.1; python_version >= '3.8'
opentelemetry-exporter-otlp-proto-common==1.33.1; python_version >= '3.8'
opentelemetry-exporter-otlp-proto-grpc==1.33.1; python_version >= '3.8'
opentelemetry-instrumentation==0.54b1; python_version >= '3.8'
opentelemetry-instrumentation-asgi==0.54b1; python_version >= '3.8'
opentelemetry-instrumentation-fastapi==0.54b1; python_version >= '3.8'
opentelemetry-proto==1.33.1; python_version >= '3.8'
opentelemetry-sdk==1.33.1; python_version >= '3.8'
opentelemetry-semantic-conventions==0.54b1; python_version >= '3.8'
opentelemetry-util-http==0.54b1; python_version >= '3.8'
orjson==3.10.18; python_version >= '3.9'
overrides==7.7.0; python_version >= '3.6'
packaging==23.2; python_version >= '3.7'
pandas==2.2.3; python_version >= '3.9'
pillow==11.2.1; python_version >= '3.9'
posthog==4.2.0
propcache==0.3.1; python_version >= '3.9'
protobuf==5.29.5; python_version >= '3.8'
pulsar-client==3.7.0
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.2; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pydantic==2.11.5; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pydub==0.25.1
pygments==2.19.1; python_version >= '3.8'
pypika==0.48.9
pyproject-hooks==1.2.0; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.1.0; python_version >= '3.9'
python-multipart==0.0.20; python_version >= '3.8'
pytz==2025.2
pyyaml==6.0.2; python_version >= '3.8'
regex==2024.11.6; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
rich==14.0.0; python_full_version >= '3.8.0'
rsa==4.9.1; python_version >= '3.6' and python_version < '4'
ruff==0.11.12; sys_platform != 'emscripten'
safehttpx==0.1.6; python_version >= '3.10'
semantic-version==2.10.0; python_version >= '2.7'
shellingham==1.5.4; python_version >= '3.7'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.41; python_version >= '3.7'
starlette==0.46.2; sys_platform != 'emscripten'
sympy==1.14.0; python_version >= '3.9'
tenacity==8.5.0; python_version >= '3.8'
tiktoken==0.9.0; python_version >= '3.9'
tokenizers==0.21.1; python_version >= '3.9'
tomlkit==0.13.2; python_version >= '3.8'
tqdm==4.67.1; python_version >= '3.7'
typer==0.16.0; python_version >= '3.7'
typing-extensions==4.13.2; python_version >= '3.8'
typing-inspect==0.9.0
typing-inspection==0.4.1; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
urllib3==2.4.0; python_version >= '3.9'
uvicorn[standard]==0.34.2; python_version >= '3.9'
uvloop==0.21.0
watchdog==6.0.0; python_version >= '3.9'
watchfiles==1.0.5
websocket-client==1.8.0; python_version >= '3.8'
websockets==15.0.1; python_version >= '3.9'
wrapt==1.17.2; python_version >= '3.8'
yarl==1.20.0; python_version >= '3.9'
zipp==3.22.0; python_version >= '3.9'